+++ mlir/lib/Bindings/Python/IRTypes.cpp
@@ -7,14 +7,13 @@
 //===----------------------------------------------------------------------===//
 
 // clang-format off
-#include "IRModule.h"
+#include "mlir/Bindings/Python/IRCore.h"
 #include "mlir/Bindings/Python/IRTypes.h"
 // clang-format on
 
 #include <optional>
 
-#include "IRModule.h"
-#include "NanobindUtils.h"
+#include "mlir/Bindings/Python/NanobindUtils.h"
 #include "mlir-c/BuiltinAttributes.h"
 #include "mlir-c/BuiltinTypes.h"
 #include "mlir-c/Support.h"
@@ -1144,7 +1143,8 @@ public:
 
 } // namespace
 
-void mlir::python::populateIRTypes(nb::module_ &m) {
+namespace mlir::python {
+void populateIRTypes(nb::module_ &m) {
   PyIntegerType::bind(m);
   PyFloatType::bind(m);
   PyIndexType::bind(m);
@@ -1175,4 +1175,18 @@ void mlir::python::populateIRTypes(nb::module_ &m) {
   PyTupleType::bind(m);
   PyFunctionType::bind(m);
   PyOpaqueType::bind(m);
+  nb::register_exception_translator([](const std::exception_ptr &p,
+                                     void *payload) {
+  // We can't define exceptions with custom fields through pybind, so
+  // instead the exception class is defined in python and imported here.
+  try {
+    if (p)
+      std::rethrow_exception(p);
+  } catch (const MLIRError &e) {
+    nb::object obj = nb::module_::import_(MAKE_MLIR_PYTHON_QUALNAME("ir"))
+                         .attr("MLIRError")(e.message, e.errorDiagnostics);
+    PyErr_SetObject(PyExc_Exception, obj.ptr());
+  }
+});
+}
 }
