+++ mlir/python/CMakeLists.txt
@@ -532,6 +532,10 @@ declare_mlir_python_extension(MLIRPythonExtension.Core
   ROOT_DIR "${PYTHON_SOURCE_DIR}"
   SOURCES
     MainModule.cpp
+    IRAffine.cpp
+    IRAttributes.cpp
+    IRInterfaces.cpp
+    IRTypes.cpp
     Pass.cpp
     Rewrite.cpp
 
@@ -991,12 +995,8 @@ get_property(NB_LIBRARY_TARGET_NAME TARGET MLIRPythonModules.extension._mlir.dso
 list(GET NB_LIBRARY_TARGET_NAME 0 NB_LIBRARY_TARGET_NAME)
 add_mlir_library_install(${NB_LIBRARY_TARGET_NAME})
 add_mlir_library(MLIRPythonSupport
-  ${PYTHON_SOURCE_DIR}/Globals.cpp
-  ${PYTHON_SOURCE_DIR}/IRAffine.cpp
-  ${PYTHON_SOURCE_DIR}/IRAttributes.cpp
   ${PYTHON_SOURCE_DIR}/IRCore.cpp
-  ${PYTHON_SOURCE_DIR}/IRInterfaces.cpp
-  ${PYTHON_SOURCE_DIR}/IRTypes.cpp
+  ${PYTHON_SOURCE_DIR}/Globals.cpp
   EXCLUDE_FROM_LIBMLIR
   SHARED
   LINK_COMPONENTS
@@ -1008,6 +1008,13 @@ add_mlir_library(MLIRPythonSupport
 target_link_libraries(MLIRPythonSupport PUBLIC ${NB_LIBRARY_TARGET_NAME})
 nanobind_link_options(MLIRPythonSupport)
 set_target_properties(MLIRPythonSupport PROPERTIES
+  LIBRARY_OUTPUT_DIRECTORY "${MLIRPythonModules_ROOT_PREFIX}/_mlir_libs"
+  BINARY_OUTPUT_DIRECTORY "${MLIRPythonModules_ROOT_PREFIX}/_mlir_libs"
+  # Needed for windows (and doesn't hurt others).
+  RUNTIME_OUTPUT_DIRECTORY "${MLIRPythonModules_ROOT_PREFIX}/_mlir_libs"
+  ARCHIVE_OUTPUT_DIRECTORY "${MLIRPythonModules_ROOT_PREFIX}/_mlir_libs"
+)
+set_target_properties(${NB_LIBRARY_TARGET_NAME} PROPERTIES
   LIBRARY_OUTPUT_DIRECTORY "${MLIRPythonModules_ROOT_PREFIX}/_mlir_libs"
   BINARY_OUTPUT_DIRECTORY "${MLIRPythonModules_ROOT_PREFIX}/_mlir_libs"
   # Needed for windows (and doesn't hurt others).
@@ -1035,4 +1042,4 @@ endif()
 target_link_libraries(
   MLIRPythonModules.extension._mlir.dso
   PUBLIC MLIRPythonSupport)
-
+target_compile_definitions(MLIRPythonSupport PRIVATE NB_DOMAIN=${MLIR_BINDINGS_PYTHON_NB_DOMAIN})
